% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benfordFirstTwoDigit.R
\name{benfordFirstTwoDigit}
\alias{benfordFirstTwoDigit}
\title{Benford's First Two Digit Test}
\usage{
benfordFirstTwoDigit(columnToTest, returnValue = "dataframe")
}
\arguments{
\item{columnToTest}{The column of a dataframe with the values to apply Benford's Law.}

\item{returnValue}{The output type returned by the function. Options are dataframe or list.}
}
\value{
Returns a dataframe or list, depending on the argument given
}
\description{
This function performs Benford's first two digit test, using a z-critical value
of 1.96 by default (equivalent to a 95\% confidence interval). The results show each digit's
test results, as well as the overall mean absolute deviation (MAD).
}
\examples{
benfordFirstTwoDigit(iris$Sepal.Width, returnValue='list')
}
