% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotF2D.R
\name{plotF2D}
\alias{plotF2D}
\title{Plot of Benford's First Two Digit Test}
\usage{
plotF2D(firstTwoDigitdf)
}
\arguments{
\item{firstTwoDigitdf}{The results of the \code{benfordFirstTwoDigit} function in dataframe format.}
}
\value{
Returns a plot
}
\description{
This function plots Benford's first two digit test results as expected vs observed
digit frequencies.
}
\examples{
plotF2D(benfordFirstTwoDigit(iris$Sepal.Width))
}
