% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFD.R
\name{plotFD}
\alias{plotFD}
\title{Plot of Benford's First Digit Test}
\usage{
plotFD(firstDigitdf)
}
\arguments{
\item{firstDigitdf}{The results of the \code{benfordFirstDigit} function in dataframe format.}
}
\value{
Returns a plot
}
\description{
This function plots Benford's first digit test results as expected vs observed
digit frequencies.
}
\examples{
plotFD(benfordFirstDigit(iris$Sepal.Width))
}

