% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotL2D.R
\name{plotL2D}
\alias{plotL2D}
\title{Plot of Benford's Last Two Digit Test}
\usage{
plotL2D(lastTwoDigitdf)
}
\arguments{
\item{lastTwoDigitdf}{The results of the \code{benfordLastTwoDigit} function in dataframe format.}
}
\value{
Returns a plot
}
\description{
This function plots Benford's last two digit test results as expected vs observed
digit frequencies.
}
\examples{
plotL2D(benfordLastTwoDigit(iris$Sepal.Width))
}
